requirements.txt
numpy
opencv_python
Pillow
Wand